fix: prompt logic and types
