+++ src/utils/parse-arguments-into-options.ts
@@ -3,7 +3,7 @@ import arg from "arg";
 import * as https from "https";
 import { getDataFromExternalExtensionArgument } from "./external-extensions";
 import chalk from "chalk";
-import { CURATED_EXTENSIONS } from "../config";
+import { CURATED_EXTENSIONS } from "../curated-extensions";
 import { SOLIDITY_FRAMEWORKS } from "./consts";
 import { validateFoundryUp } from "./system-validation";
 import fs from "fs";
