+++ src/utils/prompt-for-missing-options.ts
@@ -1,18 +1,17 @@
-import { config } from "../config";
-import { Options, RawOptions, isDefined } from "../types";
+import { Options, RawOptions } from "../types";
 import inquirer from "inquirer";
+import { SOLIDITY_FRAMEWORKS } from "./consts";
 
 // default values for unspecified args
 const defaultOptions: RawOptions = {
   project: "my-dapp-example",
+  solidityFramework: null,
   install: true,
   dev: false,
   externalExtension: null,
   help: false,
-  solidityFramework: null,
 };
 
-const invalidQuestionNames = ["project", "install"];
 const nullExtensionChoice = {
   name: "none",
   value: null,
@@ -20,43 +19,28 @@ const nullExtensionChoice = {
 
 export async function promptForMissingOptions(options: RawOptions): Promise<Options> {
   const cliAnswers = Object.fromEntries(Object.entries(options).filter(([, value]) => value !== null));
-  const questions = [];
-
-  questions.push({
-    type: "input",
-    name: "project",
-    message: "Your project name:",
-    default: defaultOptions.project,
-    validate: (value: string) => value.length > 0,
-  });
-
-  config.questions.forEach(question => {
-    if (invalidQuestionNames.includes(question.name)) {
-      throw new Error(
-        `The name of the question can't be "${question.name}". The invalid names are: ${invalidQuestionNames
-          .map(w => `"${w}"`)
-          .join(", ")}`,
-      );
-    }
-
-    const extensions = question.extensions.filter(isDefined);
-
-    const hasNoneOption = question.extensions.includes(null);
-
-    questions.push({
-      type: question.type === "multi-select" ? "checkbox" : "list",
-      name: question.name,
-      message: question.message,
-      choices: hasNoneOption ? [...extensions, nullExtensionChoice] : extensions,
-    });
-  });
-
-  questions.push({
-    type: "confirm",
-    name: "install",
-    message: "Install packages?",
-    default: defaultOptions.install,
-  });
+  const questions = [
+    {
+      type: "input",
+      name: "project",
+      message: "Your project name:",
+      default: defaultOptions.project,
+      validate: (value: string) => value.length > 0,
+    },
+    {
+      type: "list",
+      name: "solidityFramework",
+      message: "What solidity framework do you want to use?",
+      choices: [SOLIDITY_FRAMEWORKS.HARDHAT, SOLIDITY_FRAMEWORKS.FOUNDRY, nullExtensionChoice],
+      default: SOLIDITY_FRAMEWORKS.HARDHAT,
+    },
+    {
+      type: "confirm",
+      name: "install",
+      message: "Install packages?",
+      default: defaultOptions.install,
+    },
+  ];
 
   const answers = await inquirer.prompt(questions, cliAnswers);
